pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "x-infra-kit"
version = "1.0.0"
description = "Reusable CDKTF Infrastructure Library for Google Cloud Platform"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "xofyy"}
]
keywords = [
    "cdktf",
    "terraform",
    "gcp",
    "google-cloud",
    "infrastructure",
    "iac",
    "gke",
    "kubernetes"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "cdktf>=0.20.0",
    "cdktf-cdktf-provider-google>=13.0.0",
    "constructs>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/xofyy/x-infra-kit"
Documentation = "https://github.com/xofyy/x-infra-kit#readme"
Repository = "https://github.com/xofyy/x-infra-kit.git"
Issues = "https://github.com/xofyy/x-infra-kit/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["infrastructure_lib*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
